requirements.txt
sqlparse==0.4.4
pydantic>=2.7.0
langchain-core==0.2.0
langchain==0.2.0
langchain-community==0.2.0
langchain-groq>=0.1.3
pydantic>=2.7.0
streamlit>=1.34.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0